# tareaApi/tareaApi/swagger.yaml
openapi: 3.0.0
info:
  title: Tareas API
  version: 1.0.0
tags:
  - name: Tareas
paths:
  /tareas:
    get:
      tags:
          - Tareas
      summary: Obtiene todos las Tareas
      responses:
        '200':
          description: Lista de todas las Tareas
          content:
            application/json:
              example:
                - id: 1
                  name: Arreglar pc
                  type: Informatica
                - id: 2
                  name: Limpiar mesa
                  type: Limpieza
    post:
      tags:
        - Tareas  
      summary: Agrega una nueva tarea
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaInput'
      responses:
        '201':
          description: Tarea agregada exitosamente a la lista
  /tareas/{id}:
    put:
      tags:
        - Tareas  
      summary: Actualiza una tarea existente en la lista
      parameters:
        - in: path
          name: id
          required: true
          description: ID de la tarea a actualizar
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaInput'
      responses:
        '200':
          description: Tarea actualizada exitosamente
        '404':
          description: Tarea no encontrada
    delete:
      tags:
        - Tareas  
      summary: Elimina una tarea de la lista
      parameters:
        - in: path
          name: id
          required: true
          description: ID de la tarea a eliminar
          schema:
            type: integer
      responses:
        '200':
          description: Tarea eliminada exitosamente
        '404':
          description: Tarea no encontrada
components:
  schemas:
    TareaInput:
      type: object
      properties:
        name:
          type: string
          description: Nombre de la tarea
        type:
          type: string
          description: Tipo de la tarea